Decompose rust_admin media.rs into cohesive submodules
media/ now holds probe (ffprobe integration), transcode (ffmpeg
invocation and rendition orchestration), encoding (bitrate/dimension
math and validation), segments, and paths (workspace containment).
Pure move: private items become pub(crate) for the module-root glob
re-exports; call sites and tests unchanged.
